+++ NEWS.md
@@ -1,7 +1,7 @@
 # eph 0.2.0
 * enhace get_microdata. Now downloads multiple datasets
 * add skip_if_offline and some skip_on_cran for time-consuming tests
-
+* add stop if there is no internet connection for get_microdata
 
 # eph 0.1.1